services/api/pyproject.toml
[project]
name = "dprk-cti-api"
version = "0.1.0"
description = "FastAPI scaffold for DPRK CTI platform"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "authlib>=1.4.0",
  "cachetools>=5.0",
  "fastapi>=0.115.0",
  "httpx>=0.27",
  "itsdangerous>=2.2",
  "psycopg[binary]>=3.2.3",
  "pydantic-settings>=2.6.1",
  "redis>=5",
  "sqlalchemy[asyncio]>=2.0.36",
  "uvicorn[standard]>=0.32.0",
  "opentelemetry-api>=1.27",
  "opentelemetry-sdk>=1.27",
  "opentelemetry-exporter-otlp-proto-grpc>=1.27",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
  "opentelemetry-instrumentation-httpx>=0.48b0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5",
    "fakeredis>=2.23",
    "respx>=0.21",
    "freezegun>=1.5",
    "cryptography>=42",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-ra", "--strict-markers", "--strict-config", "--rootdir=services/api"]
filterwarnings = [
    "ignore::DeprecationWarning:authlib.*",
    "ignore::DeprecationWarning:fakeredis.*",
]

[tool.uv]
package = true
